+++ src/services/lampooProduct.ts
@@ -1,0 +1,9 @@
+export const LampooProductDetailsScraperObject = {
+    async findLampooProductDetails({ urls, browserInstance }: any) {
+        try {
+
+        } catch (error) {
+            console.log(error)
+        }
+    }
+};
