+++ src/services/luxuryProduct.ts
@@ -1,0 +1,9 @@
+export const LuxuryProductDetailsScraperObject = {
+    async findLuxuryProductDetails({ urls, browserInstance }: any) {
+        try {
+
+        } catch (error) {
+            console.log(error)
+        }
+    }
+};
