+++ src/services/theRealProduct.ts
@@ -1,0 +1,9 @@
+export const RealProductDetailsScraperObject = {
+    async findRealProductDetails({ urls, browserInstance }: any) {
+        try {
+
+        } catch (error) {
+            console.log(error)
+        }
+    }
+};
