+++ src/services/thredupProduct.ts
@@ -1,0 +1,9 @@
+export const ThredupProductDetailsScraperObject = {
+    async findThredupProductDetails({ urls, browserInstance }: any) {
+        try {
+
+        } catch (error) {
+            console.log(error)
+        }
+    }
+};
